Rename post info main and split views
