+++ rodolphe/main/admin.py
@@ -2,5 +2,5 @@ from django.contrib import admin
 
 # Register your models here.
 
-from post.models import Post
+from main.models import Post
 admin.site.register(Post)
