+++ rodolphe/main/forms.py
@@ -2,7 +2,7 @@ from django import forms
 from django.conf import settings
 from django.utils.translation import ugettext as _
 
-from post.models import Post, Tag
+from main.models import Post, Tag
 from captcha.fields import CaptchaField
