+++ rodolphe/main/views/__init__.py
@@ -1,0 +1,3 @@
+from main.views.paging import page
+
+home = page
