+++ rodolphe/templates/delete.html
@@ -6,7 +6,7 @@
 
 {% block content %}
 <p>
-  <a href="{% url 'post.views.view' post.thread.id %}" class="btn btn-default pull-right">
+  <a href="{% url 'main.views.thread.view' post.thread.id %}" class="btn btn-default pull-right">
     {% bootstrap_icon "hand-left" %} {% trans "back" %}
   </a>
   <h2>{% trans "deleting post" %} {{post}}</h2>
