+++ rodolphe/templates/edit.html
@@ -6,7 +6,7 @@
 
 {% block content %}
 <p>
-  <a href="{% url 'post.views.view' post.thread.id %}" class="btn btn-default pull-right">
+  <a href="{% url 'main.views.thread.view' post.thread.id %}#p{{post.id}}" class="btn btn-default pull-right">
     {% bootstrap_icon "hand-left" %} {% trans "back" %}
   </a>
   <h2>{% trans "editing post" %} {{post}}</h2>
@@ -14,7 +14,7 @@
 <hr/>
 
 <form method="post" enctype="multipart/form-data">{% csrf_token %}
-  <h3><a href="{% url 'post.views.markdown' %}">{% trans "markdown tags" %}</a></h3>
+  <h3><a href="{% url 'main.views.about.markdown' %}">{% trans "markdown tags" %}</a></h3>
   {% bootstrap_form form %}
   {% buttons %}
   <button type="submit" class="btn btn-primary">
