+++ rodolphe/templates/tag.html
@@ -1,1 +1,1 @@
-<a href="{% url 'post.views.tagsearch' tag %}" class="label label-default">{{tag}}</a>
+<a href="{% url 'main.views.tag.search' tag %}" class="label label-default">{{tag}}</a>
